Refactor TxHistory/TxRow into logic and presentation components
